don't use reflection when not necessary
