+++ src/main/scala/pl/edu/agh/hubert/engine/Input.scala
@@ -1,0 +1,8 @@
+package pl.edu.agh.hubert.engine
+
+/**
+  * Created by koperek on 30/04/16.
+  */
+class Input {
+
+}
