--- src/main/scala/pl/edu/agh/hubert/engine/engine.scala
@@ -1,5 +1,0 @@
-package pl.edu.agh.hubert
-
-package object engine {
-  type Input = Array[Array[Double]]
-}
