+++ src/test/scala/pl/edu/agh/hubert/languages/math/DifferentiationFitnessFunctionFormulaTest.scala
@@ -1,12 +1,14 @@
 package pl.edu.agh.hubert.languages.math
 
 import org.scalatest.FunSuite
+import pl.edu.agh.hubert.datasets.CSVLoader
 import pl.edu.agh.hubert.engine.{Languages, Experiment}
 import pl.edu.agh.hubert.testfixtures._
 
-class DifferentiationFitnessFunctionTest extends FunSuite {
+class DifferentiationFitnessFunctionFormulaTest extends FunSuite {
 
-  val fitnessFunction = new DifferentiationFitnessFunction(circleExperiment)
+  val fitnessFunction = new DifferentiationFitnessFunctionFormula()
+  val loadedDataSet = CSVLoader.load(dataSet("circle.csv", Set("x", "y")))
 
   test("x^2 + y^2 - 4 should have very low error") {
 
@@ -14,9 +16,7 @@ class DifferentiationFitnessFunctionTest extends FunSuite {
     val y = new Variable(1)
     val properSolution = new MathIndividual(new Minus(new Plus(new Mul(x, x), new Mul(y, y)), new Constant(4.0)))
 
-    val evaluatedPopulation = fitnessFunction.evaluatePopulation(Array(properSolution))
-
-    val fitness = evaluatedPopulation(0).fitness
+    val fitness = fitnessFunction.evaluateFitnessFormula(properSolution, loadedDataSet)
 
     assert(fitness.isDefined)
     assert(Math.abs(fitness.get) < 3e-12)
