+++ src/test/scala/pl/edu/agh/hubert/languages/math/DifferentiationWithFitnessPredictionFitnessFunctionTest.scala
@@ -1,0 +1,7 @@
+package pl.edu.agh.hubert.languages.math
+
+import org.scalatest.FunSuite
+
+class DifferentiationWithFitnessPredictionFitnessFunctionTest extends FunSuite {
+
+}
